adds usb frag when usb is detected
