+++ BundleTransport/app/src/main/java/net/discdd/bundletransport/UsbFragment.java
@@ -44,13 +44,12 @@ import java.util.stream.Stream;
 import static java.util.logging.Level.WARNING;
 
 public class UsbFragment extends Fragment {
-    private static final String USB_DIR_NAME = "DDD_transport";
-    private static final String RELATIVE_CLIENT_PATH = "client";
-    private static final String RELATIVE_SERVER_PATH = "server";
+    private UsbFileManager usbFileManager;
     private TransportPaths transportPaths;
     private Button usbExchangeButton;
     private TextView usbConnectionText;
     private UsbManager usbManager;
+/*Declaring a broadcast receiver*/
     private BroadcastReceiver mUsbReceiver;
     public static boolean usbConnected = false;
     private StorageManager storageManager;
@@ -69,14 +68,16 @@ public class UsbFragment extends Fragment {
         usbExchangeButton = view.findViewById(R.id.usb_exchange_button);
         usbConnectionText = view.findViewById(R.id.usbconnection_response_text);
 
+        storageManager = (StorageManager) getActivity().getSystemService(Context.STORAGE_SERVICE);
         usbManager = (UsbManager) getActivity().getSystemService(Context.USB_SERVICE);
+        usbFileManager = new UsbFileManager(storageManager, transportPaths);
+/*Instantiating a broadcast receiver*/
         mUsbReceiver = new BroadcastReceiver() {
             @Override
             public void onReceive(Context context, Intent intent) {
                 handleUsbBroadcast(intent);
             }
         };
-        storageManager = (StorageManager) getActivity().getSystemService(Context.STORAGE_SERVICE);
 
         //Register USB broadcast receiver
         IntentFilter filter = new IntentFilter();
@@ -90,7 +91,7 @@ public class UsbFragment extends Fragment {
         usbExchangeButton.setOnClickListener(v -> {
             logger.log(INFO, "Sync button was hit");
             try {
-                populateUsb();
+                usbFileManager.populateUsb();
             } catch (IOException e) {
                 logger.log(INFO, "Populate USB was unsuccessful");
                 throw new RuntimeException(e);
@@ -99,131 +100,10 @@ public class UsbFragment extends Fragment {
         return view;
     }
 
-    /**
-     * Finds appropriate usb dir for ddd activities and copies device files.
-     *
-     * @throws IOException
-     */
-    private void populateUsb() throws IOException {
-        List<StorageVolume> storageVolumes = storageManager.getStorageVolumes();
-        for (StorageVolume volume : storageVolumes) {
-            if (volume.isRemovable() && !volume.isEmulated()) {
-                File usbStorageDir = volume.getDirectory();
-                if (usbStorageDir != null) {
-                    File usbTransportDir = new File(usbStorageDir, USB_DIR_NAME);
-                    File usbTransportToServerDir = new File(usbTransportDir, RELATIVE_SERVER_PATH);
-                    File usbTransportToClientDir = new File(usbTransportDir, RELATIVE_CLIENT_PATH);
-                    if (!usbTransportDir.exists()) {
-                        usbTransportDir.mkdirs();
-                        usbTransportToServerDir.mkdirs();
-                        usbTransportToClientDir.mkdirs();
-                    }
-                    try {
-                        toClientList(usbTransportToClientDir);
-                        toServerList(usbTransportToServerDir);
-                    } catch (Exception e) {
-                        logger.log(WARNING, "failed to populate USB and or Android device");
-                        throw new RuntimeException("Bad call to populate USB or Android device", e);
-                    }
-                    reduceUsbFiles(usbTransportToClientDir, usbTransportToServerDir);
-                }
-                updateUsbStatus(true, "Sync successful", Color.GREEN);
-                break;
-            }
-        }
-    }
-
-    /**
-     * Copies for-client files from transport device (handling downstream).
-     *
-     * @param targetDir target directory; USBs client directory
-     * @throws IOException
-     */
-    private void toClientList(File targetDir) throws IOException, GeneralSecurityException,
-            RoutingExceptions.ClientMetaDataFileException, InvalidKeyException {
-        List<Path> storageList;
-        Path devicePathForClient = transportPaths.toClientPath;
-        try (Stream<Path> walk = Files.walk(devicePathForClient)) {
-            storageList = walk.filter(Files::isRegularFile).collect(Collectors.toList());
-        }
-        if (storageList.isEmpty()) {
-            logger.log(INFO, "No bundles to download from device to USB (to client files)");
-            logger.log(INFO, "Our empty storageList was " + devicePathForClient);
-            return;
-        }
-        for (Path deviceFilePath : storageList) {
-            Path targetPath = targetDir.toPath().resolve(deviceFilePath.getFileName());
-            Files.copy(deviceFilePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
-        }
-    }
-
-    /**
-     * Copies for-server files from usb (handling upstream).
-     *
-     * @param sourceDir source directory; USBs server directory
-     */
-    private void toServerList(File sourceDir) throws IOException {
-        List<Path> storageList;
-        try (Stream<Path> walk = Files.walk(sourceDir.toPath())) {
-            storageList = walk.filter(Files::isRegularFile).collect(Collectors.toList());
-        }
-        if (storageList.isEmpty()) {
-            logger.log(INFO, "No bundles to download from USB to device (to server files)");
-            logger.log(INFO, "Our empty storageList was " + sourceDir);
-            return;
-        }
-        Path devicePathForServer = transportPaths.toServerPath;
-        for (Path usbFilePath : storageList) {
-            Path targetPath = devicePathForServer.resolve(usbFilePath.getFileName());
-            Files.copy(usbFilePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
-        }
-    }
-
-    /**
-     * Delete USB toClient and toServer files when files missing from transport and already on transport respectively.
-     *
-     * @param usbTransportToClientDir
-     * @param usbTransportToServerDir
-     * @throws IOException
-     */
-    private void reduceUsbFiles(File usbTransportToClientDir, File usbTransportToServerDir) throws IOException {
-        List<Path> usbToClient;
-        try (Stream<Path> walk = Files.walk(usbTransportToClientDir.toPath())) {
-            usbToClient = walk.filter(Files::isRegularFile).collect(Collectors.toList());
-        }
-        boolean deletedClientFiles = false;
-        for (Path usbFile : usbToClient) {
-            File possibleFile = new File(String.valueOf(transportPaths.toClientPath), usbFile.getFileName().toString());
-            boolean usbFileExistsInAndroid = possibleFile.exists();
-            if (!usbFileExistsInAndroid) {
-                Files.deleteIfExists(usbFile);
-                deletedClientFiles = true;
-            }
-        }
-        String res = (deletedClientFiles) ? "Successfully deleted excess client files from USB" :
-                "No excess client files to delete from USB";
-        logger.log(INFO, res);
-        List<Path> androidToServer;
-        try (Stream<Path> walk = Files.walk(transportPaths.toServerPath)) {
-            androidToServer = walk.filter(Files::isRegularFile).collect(Collectors.toList());
-        }
-        boolean deletedServerFiles = false;
-        for (Path usbFile : androidToServer) {
-            File possibleFile = new File(usbTransportToServerDir, usbFile.getFileName().toString());
-            boolean androidFileExistsInUsb = possibleFile.exists();
-            if (androidFileExistsInUsb) {
-                Files.deleteIfExists(possibleFile.toPath());
-                deletedServerFiles = true;
-            }
-        }
-        res = (deletedServerFiles) ? "Successfully deleted excess server files from USB" :
-                "No excess server files to delete from USB";
-        logger.log(INFO, res);
-    }
-
     @Override
     public void onDestroyView() {
         super.onDestroyView();
+        //call new method to do the following from connection manager
         if (mUsbReceiver != null) {
             getActivity().unregisterReceiver(mUsbReceiver);
         }
@@ -231,21 +111,21 @@ public class UsbFragment extends Fragment {
 
     private void handleUsbBroadcast(Intent intent) {
         String action = intent.getAction();
-        if (UsbManager.ACTION_USB_DEVICE_DETACHED.equals(action)) {
+        if (UsbManager.ACTION_USB_DEVICE_DETACHED.equals(action)) { //shouldn't ever be reached
             updateUsbStatus(false, getString(R.string.no_usb_connection_detected), Color.RED);
             usbConnected = false;
             showUsbDetachedToast();
-        } else if (UsbManager.ACTION_USB_DEVICE_ATTACHED.equals(action)) {
+        } else if (UsbManager.ACTION_USB_DEVICE_ATTACHED.equals(action)) { //should always be the case (barely visible?)
             updateUsbStatus(false, getString(R.string.usb_device_attached_checking_for_storage_volumes), Color.BLUE);
             scheduledExecutor.schedule(() -> checkUsbConnection(2), 1, TimeUnit.SECONDS);
             showUsbAttachedToast();
         }
     }
 
-    private void checkUsbConnection(int tries) {
+    public boolean checkUsbConnection(int tries) {
         usbConnected = !usbManager.getDeviceList().isEmpty();
         getActivity().getMainExecutor().execute(() -> {
-            if (!usbManager.getDeviceList().isEmpty()) {
+            if (usbConnected) {
                 updateUsbStatus(true, getString(R.string.usb_connection_detected), Color.GREEN);
             } else {
                 updateUsbStatus(false, getString(R.string.no_usb_connection_detected), Color.RED);
@@ -254,9 +134,11 @@ public class UsbFragment extends Fragment {
         if (tries > 0 && !usbConnected) {
             scheduledExecutor.schedule(() -> checkUsbConnection(tries - 1), 1, TimeUnit.SECONDS);
         }
+        return usbConnected;
     }
 
     public void updateUsbStatus(boolean isConnected, String statusText, int color) {
+        //get parameter info from checkUsbConnection in connection manager
         getActivity().runOnUiThread(() -> {
             usbExchangeButton.setEnabled(isConnected);
             usbConnectionText.setText(statusText);
@@ -264,10 +146,12 @@ public class UsbFragment extends Fragment {
         });
     }
 
+    //should be deleted
     private void showUsbAttachedToast() {
         Toast.makeText(getActivity(), getString(R.string.usb_device_attached), Toast.LENGTH_SHORT).show();
     }
 
+    //should proceed with closing tab
     private void showUsbDetachedToast() {
         Toast.makeText(getActivity(), getString(R.string.usb_device_detached), Toast.LENGTH_SHORT).show();
     }
